Compiles and runs unit tests, but one of them fails

+++ src/Mqtt/Stream.hs
@@ -1,8 +1,8 @@
 module Mqtt.Stream (nextMessage, containsFullMessage,
                     mqttMessages, mqttStream) where
 
-import           Data.ByteString (pack, unpack, empty, append)
-import qualified Data.ByteString as BS
+import           Data.ByteString.Lazy (pack, unpack, empty, append)
+import qualified Data.ByteString.Lazy as BS
 import           Mqtt.Message (getRemainingLength)
 
 
@@ -19,7 +19,7 @@ containsFullMessage :: BS.ByteString -> Bool
 containsFullMessage pkt = let remaining = getRemainingLength pkt
                               headerLen = 2
                               totalLen = remaining + headerLen in
-                          remaining >= 0 && BS.length pkt >= totalLen
+                          remaining >= 0 && BS.length pkt >= fromIntegral totalLen
 
 -- Takes a bytestring and returns a list of MQTT messages plus a bytestring
 -- of remaining bytes
